refactor : Reply 를 리스트로 받을 수 있도록 구조 변경
