+++ src/main/java/com/formmaker/fff/common/type/QuestionTypeEnum.java
@@ -3,7 +3,7 @@ package com.formmaker.fff.common.type;
 
 import com.formmaker.fff.question.dto.QuestionDto;
 
-import com.formmaker.fff.reply.dto.request.ReplyDto;
+import com.formmaker.fff.reply.dto.ReplyDto;
 import com.formmaker.fff.stats.dto.QuestionStats;
 import static com.formmaker.fff.stats.StatsMethod.statsMethod;
 import lombok.Getter;
@@ -11,8 +11,6 @@ import lombok.Getter;
 import java.util.List;
 
 import java.util.function.BiFunction;
-
-import java.util.function.Function;
 
 @Getter
 public enum QuestionTypeEnum {
