+++ src/main/java/com/formmaker/fff/question/dto/request/QuestionCreateRequest.java
@@ -1,6 +1,5 @@
 package com.formmaker.fff.question.dto.request;
 
-import com.formmaker.fff.answer.dto.request.AnswerCreateRequest;
 import com.formmaker.fff.common.type.QuestionTypeEnum;
 import lombok.Getter;
 
@@ -11,7 +10,6 @@ public class QuestionCreateRequest {
     private QuestionTypeEnum questionType;
     private String summary;
     private Integer volume;
-    private Integer questionNum;
     private String questionTitle;
     private List<String> answerList;
 }
