+++ src/main/java/com/formmaker/fff/reply/controller/ReplyController.java
@@ -2,13 +2,15 @@ package com.formmaker.fff.reply.controller;
 
 import com.formmaker.fff.common.response.ResponseMessage;
 import com.formmaker.fff.common.response.security.UserDetailsImpl;
-import com.formmaker.fff.reply.dto.request.ReplyCreateRequest;
+import com.formmaker.fff.reply.dto.request.EachReply;
 import com.formmaker.fff.reply.service.ReplyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -23,7 +25,7 @@ public class ReplyController {
 //    }
 
     @PostMapping("/{surveyId}/reply")
-    public ResponseEntity<ResponseMessage> postReply(@PathVariable Long  surveyId, @RequestBody ReplyCreateRequest replyRequest, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+    public ResponseEntity<ResponseMessage> postReply(@PathVariable Long  surveyId, @RequestBody List<EachReply> replyRequest, @AuthenticationPrincipal UserDetailsImpl userDetails) {
         replyService.postReply(surveyId, replyRequest, userDetails);
         ResponseMessage responseMessage = new ResponseMessage("설문 응답 성공", 200, null);
         return new ResponseEntity<>(responseMessage, HttpStatus.valueOf(responseMessage.getStatusCode()));
