+++ src/main/java/com/formmaker/fff/reply/dto/request/ReplyCreateRequest.java
@@ -1,6 +1,5 @@
 package com.formmaker.fff.reply.dto.request;
 
-import com.formmaker.fff.common.type.QuestionTypeEnum;
 import lombok.Getter;
 
 import java.util.ArrayList;
@@ -8,9 +7,5 @@ import java.util.List;
 
 @Getter
 public class ReplyCreateRequest {
-    private Long questionId;
-    private QuestionTypeEnum questionType;
-    private Integer questionNum;
-    private List<Integer> selectValue = new ArrayList<>();
-    private String descriptive;
+    private List<EachReply> replyList = new ArrayList<>();
 }
